projects/flow.aggregation-mpi/requirements.txt
mpi4py>=3.0.0
signac>=2.0.0
signac-flow>=0.25.0
